Fixed Frame Issue
There was a bug in which if you attempt to create an additional frame the program would crash as it can only handle a limited number of frames. I did change the original C code for the first time since Jan 2017 as it caused the C code to crash so I fixed it.